feature/chat+cleanup

+++ src/main/java/thelaborseekers/jobhubapi/repository/PostulacionRepository.java
@@ -8,10 +8,10 @@ import thelaborseekers.jobhubapi.model.entity.JobOffer;
 import thelaborseekers.jobhubapi.model.entity.Postulacion;
 
 @Repository
-public interface PostulacionRepository extends JpaRepository<Postulacion, Long> {
+public interface PostulacionRepository extends JpaRepository<Postulacion, Integer> {
      // Nuevo método para obtener el historial de postulaciones de un postulante
     List<Postulacion> findByPostulanteId(Integer postulanteId);
     // Método que encuentra postulaciones según una oferta laboral
     List<Postulacion> findByOfertaLaboral(JobOffer jobOffer);
-    List<Postulacion> findByOfertaLaboralId(Long ofertaLaboralId);
+    List<Postulacion> findByOfertaLaboralId(Integer ofertaLaboralId);
 }
